pyproject.toml
[project]
name = "image1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
